pyproject.toml
[tool.poetry]
name = "snel-pointless"
version = "0.1.0"
description = "A natural language crypto command interpreter"
authors = ["papa <papaandthejimjams@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.8"
uvicorn = "0.34.0"
httpx = "0.28.1"
python-dotenv = "1.0.1"
slowapi = "0.1.9"
eth-rpc-py = "0.1.26"
eth-account = "0.13.5"
eth-typing = "5.2.0"
dowse = "^0.1.6.post1"
redis = "^5.2.1"
web3 = "^7.8.0"
upstash-redis = "^1.2.0"
pydantic = "2.10.6"
pycryptodome = "^3.21.0"
pydantic-settings = "^2.2.1"
openai = "^1.65.2"
eth-utils = "^5.2.0"
python-telegram-bot = "^22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.3"
ipython = "^8.32.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["E704"]
src-paths = ["./src"]
line-length = 120

[tool.pytest.ini_options]
pythonpath = ["src"]
filterwarnings = [
    "ignore::DeprecationWarning:passlib.*:",
    "ignore::DeprecationWarning:jose.*:",
    "ignore::DeprecationWarning:websockets.legacy.*",
    "ignore::DeprecationWarning:pytest_asyncio.*",
]
asyncio_mode = "auto"

[tool.poetry.scripts]
main = "main:main"
simple_program = "examples.program.swap.executor:main"
